feat: began 2nd scene, added shao's other animations, fix: clamped camera correctly
need to use the attack scriptable objects to define the enemy sequences in the second room.

also need to make it possible to go to the second room at all.

as well, we need to define each sequence of 3 buttons as a move (eg. ILM is punch), so whenever we mention ILM again we actually say "punch" and the player has to remember it's ILM